Topbar: Refactor to react-redux
