+++ docs/frontend.rst
@@ -204,5 +204,5 @@ General Components
 +-------------------------------+--------------------------------------------------+--------------------------+
 | ``timetable_name_input.jsx``  | .. image:: components/timetable_name_input.png   |                          |
 +-------------------------------+--------------------------------------------------+--------------------------+
-| ``top_bar.jsx``               | .. image:: components/top_bar.png                |                          |
+| ``TopBar.tsx``               | .. image:: components/top_bar.png                |                          |
 +-------------------------------+--------------------------------------------------+--------------------------+
